Move scala classes into more logical package hierarchy
