+++ app/src/main/java/com/waz/zclient/pages/main/profile/camera/CameraFragment.java
@@ -37,7 +37,7 @@ import com.waz.api.ImageAssetFactory;
 import com.waz.zclient.OnBackPressedListener;
 import com.waz.zclient.R;
 import com.waz.zclient.camera.CameraPreviewObserver;
-import com.waz.zclient.camera.CameraPreviewTextureView;
+import com.waz.zclient.camera.views.CameraPreviewTextureView;
 import com.waz.zclient.camera.FlashMode;
 import com.waz.zclient.controllers.accentcolor.AccentColorObserver;
 import com.waz.zclient.controllers.camera.CameraActionObserver;
