+++ app/src/main/res/layout/calling__participants__chathead.xml
@@ -35,7 +35,7 @@
             android:layout_height="match_parent"
             />
 
-        <com.waz.zclient.views.chathead.ChatheadView
+        <com.waz.zclient.common.views.ChatheadView
             android:id="@+id/chv__calling__participant__chathead"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
@@ -46,7 +46,7 @@
 
     </FrameLayout>
 
-    <com.waz.zclient.calling.UserNameView
+    <com.waz.zclient.calling.views.UserNameView
         android:layout_marginTop="@dimen/wire__padding__small"
         android:id="@+id/ttv__calling__participant__name"
         android:layout_width="match_parent"
