+++ app/src/main/res/layout/chathead_with_text_footer.xml
@@ -24,7 +24,7 @@
        android:layout_height="wrap_content"
     >
 
-    <com.waz.zclient.views.chathead.ChatheadView
+    <com.waz.zclient.common.views.ChatheadView
         android:id="@+id/cv__chathead"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
