+++ app/src/main/scala/com/waz/zclient/calling/controllers/CallPermissionsController.scala
@@ -15,11 +15,12 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.waz.zclient.calling
+package com.waz.zclient.calling.controllers
 
 import com.waz.api.VoiceChannelState.OTHER_CALLING
 import com.waz.model.ConvId
 import com.waz.zclient._
+import com.waz.zclient.common.controllers.{CameraPermission, RecordAudioPermission, PermissionsController}
 import timber.log.Timber
 
 /**
